Improve Code Quality (#38)
* Add logic operations between field elements

Add explicit loop for logic operations AND and XOR in feature
`trace` code.

* Update to edition 2021

* fix: improve code quality standards across the crate

- fixed clippy suggestions or marked where clippy is wrong
- fixed broken doc links
- removed `get_domain_attrs` in favor of `EvaluationDomainExt`
- general best-practices improvements

* fix: remove unnecessary import renaming

* fix: use Range instead of RangeInclusive

* fix: improve CI workflow, include clippy and docs

* chore: fix typo

* fix: resolve comments, suggestions, and some formatting

- fixed some `clippy` lints directly instead of leaving `allow` cues, left others for future refactoring
- fixed old jubjub comment
- removed `crate::error::Result` type
- remove some type declaration redundancies
- sorted `derive` macro invocations
- added warning denial to CI
- resolved other PR comments

+++ src/permutation/constants.rs
@@ -1,20 +1,24 @@
-#![allow(non_snake_case)]
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+//! Constants used in the permutation argument to ensure that the wire subsets
+//! are disjoint.
+
+#![allow(non_snake_case)]
+
 use ark_ff::PrimeField;
 
-/// Constants used in the permutation argument to ensure that the wire subsets
-/// are disjoint.
 pub(crate) fn K1<F: PrimeField>() -> F {
-    F::from(7 as u64)
+    F::from(7_u64)
 }
+
 pub(crate) fn K2<F: PrimeField>() -> F {
-    F::from(13 as u64)
+    F::from(13_u64)
 }
+
 pub(crate) fn K3<F: PrimeField>() -> F {
-    F::from(17 as u64)
+    F::from(17_u64)
 }
